backend/requirements.txt
# Core API
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.0.0

# AI/ML - Groq API
groq==0.9.0

# NLP/ML - Legal BERT (NEW)
transformers==4.35.0
torch==2.1.0
spacy==3.7.2
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.0

# Data Processing
numpy<2
pandas==2.0.3
python-dateutil==2.8.2

# HTTP Client
requests==2.31.0
httpx==0.24.1

# Utilities
aiofiles==23.1.0
